Edits, see description for info
Added/fixed overlay img for contact us page. Edited the form - changed all names of radio buttons to radio, fixes the multi radio select issue. Added ids to them and then added "for" in label to connect the two, now you can click on the text or radio button to select.
Last thing was index, edited the styling of section 2 for products.